Feat : [SHK-141] 피드 최근 날짜 순 조회 기능 구현
- (중요) 조회 성능을 높이기 위해 좋아요를 필드로 가지도록 엔티티, DDL과 관련 코드 수정
- 다른 도메인에서 수정된 스키마(Order, Transaction)까지 통합 반영
- QueryDSL 저장, 수정, 삭제 외 조회에 QueryDSL 적용

+++ src/main/java/com/prgrms/kream/domain/style/service/StyleService.java
@@ -2,16 +2,13 @@ package com.prgrms.kream.domain.style.service;
 
 import static com.prgrms.kream.common.mapper.StyleMapper.*;
 
-import java.util.List;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 import javax.persistence.EntityNotFoundException;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.prgrms.kream.domain.style.dto.request.GetFeedServiceRequest;
 import com.prgrms.kream.domain.style.dto.request.LikeFeedServiceRequest;
 import com.prgrms.kream.domain.style.dto.request.RegisterFeedServiceRequest;
 import com.prgrms.kream.domain.style.dto.request.UpdateFeedServiceRequest;
@@ -50,10 +47,13 @@ public class StyleService {
 	}
 
 	@Transactional(readOnly = true)
-	public GetFeedServiceResponses get(GetFeedServiceRequest getFeedServiceRequest) {
-		return toGetFeedServiceResponses(
-				feedRepository.findAllById(getFeedServiceRequest.ids())
-		);
+	public GetFeedServiceResponses getNewestFeeds() {
+		return toGetFeedServiceResponses(feedRepository.findAllByRecent());
+	}
+
+	@Transactional(readOnly = true)
+	public GetFeedServiceResponses getAllByTag(String tag) {
+		return toGetFeedServiceResponses(feedRepository.findAllByTag(tag));
 	}
 
 	@Transactional
@@ -92,28 +92,35 @@ public class StyleService {
 				likeFeedServiceRequest.feedId(),
 				likeFeedServiceRequest.memberId()
 		)) {
+			feedRepository.findById(likeFeedServiceRequest.feedId())
+					.map(feed -> {
+						feed.increaseLikes();
+						return feedRepository.save(feed);
+					})
+					.orElseThrow(EntityNotFoundException::new);
+
 			feedLikeRepository.save(toFeedLike(likeFeedServiceRequest));
 		}
 	}
 
 	@Transactional
 	public void deleteFeedLike(LikeFeedServiceRequest likeFeedServiceRequest) {
-		feedLikeRepository.deleteByFeedIdAndMemberId(
+		if (feedLikeRepository.existsByFeedIdAndMemberId(
 				likeFeedServiceRequest.feedId(),
 				likeFeedServiceRequest.memberId()
-		);
-	}
+		)) {
+			feedRepository.findById(likeFeedServiceRequest.feedId())
+					.map(feed -> {
+						feed.decreaseLikes();
+						return feedRepository.save(feed);
+					})
+					.orElseThrow(EntityNotFoundException::new);
 
-	@Transactional(readOnly = true)
-	public List<Long> getAllByTag(String tag) {
-		return feedTagRepository.findAllByTag(tag).stream()
-				.map(FeedTag::getFeedId)
-				.collect(Collectors.toList());
-	}
-
-	@Transactional(readOnly = true)
-	public Long getFeedLike(Long id) {
-		return feedLikeRepository.countByFeedId(id);
+			feedLikeRepository.deleteByFeedIdAndMemberId(
+					likeFeedServiceRequest.feedId(),
+					likeFeedServiceRequest.memberId()
+			);
+		}
 	}
 
 }
